Querying node measures

These functions are a collection of node measures that do not really fall into the class of centrality measures. For lack of a better place they are collected under the node_* umbrella of functions.

node_eccentricity(mode = "out")

node_constraint(weights = NULL)

node_coreness(mode = "out")

node_diversity(weights)

node_bridging_score()

node_effective_network_size()

node_connectivity_impact()

node_closeness_impact()

node_fareness_impact()

Arguments

mode

The way edges should be followed in the case of directed graphs.

weights

The weights to use for each node during calculation

Value

A numeric vector of the same length as the number of nodes in the graph.

Functions

node_eccentricity(): measure the maximum shortest path to all other nodes in the graph

node_constraint(): measures Burts constraint of the node. See igraph::constraint()

node_coreness(): measures the coreness of each node. See igraph::coreness()

node_diversity(): measures the diversity of the node. See igraph::diversity()

node_bridging_score(): measures Valente's Bridging measures for detecting structural bridges (influenceR)

node_effective_network_size(): measures Burt's Effective Network Size indicating access to structural holes in the network (influenceR)

node_connectivity_impact(): measures the impact on connectivity when removing the node (NetSwan)

node_closeness_impact(): measures the impact on closeness when removing the node (NetSwan)

node_fareness_impact(): measures the impact on fareness (distance between all node pairs) when removing the node (NetSwan)

Examples

# Calculate Burt's Constraint for each node
create_notable('meredith') %>%
  mutate(b_constraint = node_constraint())
#> # A tbl_graph: 70 nodes and 140 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 70 × 1 (active)
#>   b_constraint
#>          <dbl>
#> 1         0.25
#> 2         0.25
#> 3         0.25
#> 4         0.25
#> 5         0.25
#> 6         0.25
#> # … with 64 more rows
#> #
#> # Edge Data: 140 × 2
#>    from    to
#>   <int> <int>
#> 1     1     5
#> 2     1     6
#> 3     1     7
#> # … with 137 more rows