Calculate node pair properties

This set of functions can be used for calculations that involve node pairs. If the calculateable measure is not symmetric the function will come in two flavours, differentiated with _to/_from suffix. The *_to() functions will take the provided node indexes as the target node (recycling if necessary). For the *_from() functions the provided nodes are taken as the source. As for the other wrappers provided, they are intended for use inside the tidygraph framework and it is thus not necessary to supply the graph being computed on as the context is known.

node_adhesion_to(nodes)

node_adhesion_from(nodes)

node_cohesion_to(nodes)

node_cohesion_from(nodes)

node_distance_to(nodes, mode = "out", weights = NULL, algorithm = "automatic")

node_distance_from(
  nodes,
  mode = "out",
  weights = NULL,
  algorithm = "automatic"
)

node_cocitation_with(nodes)

node_bibcoupling_with(nodes)

node_similarity_with(nodes, mode = "out", loops = FALSE, method = "jaccard")

node_max_flow_to(nodes, capacity = NULL)

node_max_flow_from(nodes, capacity = NULL)

Arguments

nodes

The other part of the node pair (the first part is the node defined by the row). Recycled if necessary.

mode

How should edges be followed? If 'all' all edges are considered, if 'in' only inbound edges are considered, and if 'out' only outbound edges are considered

weights

The weights to use for calculation

algorithm

The distance algorithms to use. By default it will try to select the fastest suitable algorithm. Possible values are "automatic", "unweighted", "dijkstra", "bellman-ford", and "johnson"

loops

Should loop edges be considered

method

The similarity measure to calculate. Possible values are: "jaccard", "dice", and "invlogweighted"

capacity

The edge capacity to use

Value

A numeric vector of the same length as the number of nodes in the graph

Functions

node_adhesion_to(): Calculate the adhesion to the specified node. Wraps igraph::edge_connectivity()

node_adhesion_from(): Calculate the adhesion from the specified node. Wraps igraph::edge_connectivity()

node_cohesion_to(): Calculate the cohesion to the specified node. Wraps igraph::vertex_connectivity()

node_cohesion_from(): Calculate the cohesion from the specified node. Wraps igraph::vertex_connectivity()

node_distance_to(): Calculate various distance metrics between node pairs. Wraps igraph::distances()

node_distance_from(): Calculate various distance metrics between node pairs. Wraps igraph::distances()

node_cocitation_with(): Calculate node pair cocitation count. Wraps igraph::cocitation()

node_bibcoupling_with(): Calculate node pair bibliographic coupling. Wraps igraph::bibcoupling()

node_similarity_with(): Calculate various node pair similarity measures. Wraps igraph::similarity()

node_max_flow_to(): Calculate the maximum flow to a node. Wraps igraph::max_flow()

node_max_flow_from(): Calculate the maximum flow from a node. Wraps igraph::max_flow()

Examples

# Calculate the distance to the center node
create_notable('meredith') %>%
  mutate(dist_to_center = node_distance_to(node_is_center()))
#> # A tbl_graph: 70 nodes and 140 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 70 × 1 (active)
#>   dist_to_center
#>            <dbl>
#> 1              1
#> 2              1
#> 3              1
#> 4              5
#> 5              6
#> 6              6
#> # … with 64 more rows
#> #
#> # Edge Data: 140 × 2
#>    from    to
#>   <int> <int>
#> 1     1     5
#> 2     1     6
#> 3     1     7
#> # … with 137 more rows